requirements.txt
aiosmtplib==2.0.1
alembic==1.9.0
anyio==3.6.2
asyncclick==8.1.3.4
asyncpg==0.27.0
bcrypt==4.0.1
blinker==1.5
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
databases==0.6.2
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.88.0
fastapi-mail==1.2.5
greenlet==2.0.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
python-decouple==3.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
python-dotenv==0.21.1